Set all repl as "manual"
This is part of an attempt to improve situation in #1598.

When calling `bazel build //...`, repl runfiles directory were created,
generating a lot of files on the disk. This was responsible for
important slowdown as well as disk usage.

This change set the `"manual"` tag to all implicit repl targets, so they
are only built (and run) when using explicit bazel call.